コミット漏れ

+++ KelpNet/Common/Real.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-//using RealType = System.Double;
-using RealType = System.Single;
+using RealType = System.Double;
+//using RealType = System.Single;
 
 namespace KelpNet.Common
 {
